requirements_dev.txt
alembic==1.4.3
astroid==2.5.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
isort==5.8.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.6.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.812
mypy-extensions==0.4.3
openpyxl==3.0.5
packaging==20.4
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pylint==2.7.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
